- Improved implementation of SVC detection in liblts - Removed gcc 4 warnings from ltsmin
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1343 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/ltsmindata.cpp
@@ -514,7 +514,7 @@ static void TransitionsGoingToBlock(SVCint b, ATermList *newlab) {
    ATermList newlabels = ATempty;
    GetBlockBoundaries(b, &left, &right);
    if (classes) fprintf(stdout,
-   "--------------------------- block %d --------------------------\n", b);
+   "--------------------------- block %d --------------------------\n", (int) b);
    /* ATwarning("TransitionGoingTo b = %d newb = %d\n",b, newb); */
    for (i = left; i < right;i++) {
         ATermList labels = lab[s[i]], pars = ATempty;
